requirements.txt
langgraph>=0.4.0
pydantic-ai>=0.2.0
pydantic>=2.0.0
langchain-anthropic>=0.3.0
langchain-ollama>=0.3.0
chromadb>=0.6.0
tenacity>=9.0.0
tiktoken>=0.9.0
typer>=0.15.0
rich>=13.0.0
structlog>=25.0.0
python-dotenv>=1.0.0
langgraph-checkpoint-sqlite>=2.0.0

# Dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.25.0
mypy>=1.0.0
ruff>=0.9.0
